requirements.txt
black==21.10b0
Django==3.0.5
django-extensions==3.1.3
django-filter==2.2.0
django-graphql-jwt==0.3.4
graphene==2.1.8
graphene-django==2.9.1
psycopg2==2.8.6
